Vectorization: split into modules
