+++ orangecontrib/text/vectorization/__init__.py
@@ -1,0 +1,3 @@
+from .bagofwords import CountVectorizer
+from .tfidf import TfidfVectorizer
+from .simhash import SimhashVectorizer
